refactored

+++ main.cpp
@@ -1,5 +1,3 @@
-#include <bits/stdc++.h>
-#include "LexicalAnalyzer/lexicalAnalyzer.h"
 #include "patternmatcher/PatternMatcher.h"
 #include "NFA/NFA.h"
 #include "DFA/DFA.h"
@@ -9,13 +7,14 @@
 using namespace std;
 
 unordered_map<string, int>
-getPriority(vector<pair<string, vector<string> > > exprs, const vector<string> &keywords,
+getPriority(vector<pair<string, vector<string>>> exprs, const vector<string> &keywords,
             const vector<string> &punctuations);
 
 NFA convert_exprs_postfix_to_NFA(const vector<pair<string, vector<string>>> &exprs, const vector<string> &keywords,
                                  const vector<string> &punctuations);
 
-int main() {
+int main()
+{
     string projectPath = R"(D:\Development Workshop\Syntax-Directed-Translator\)";
     string outPath = projectPath + R"(Outputs\)";
 
@@ -27,11 +26,11 @@ int main() {
     vector<string> punctuations_lines = Parser::get_punctuation_lines(rules);
     vector<string> punctuations = Parser::parse_keywords(punctuations_lines);
 
-    vector<pair<string, string> > defs_lines = Parser::get_regular_def_lines(rules);
-    unordered_map<string, vector<string> > defs = Parser::parse_defs(defs_lines);
+    vector<pair<string, string>> defs_lines = Parser::get_regular_def_lines(rules);
+    unordered_map<string, vector<string>> defs = Parser::parse_defs(defs_lines);
 
-    vector<pair<string, string> > expr_lines = Parser::get_regular_expr_lines(rules);
-    vector<pair<string, vector<string> > > exprs = Parser::parse_expr(expr_lines, defs);
+    vector<pair<string, string>> expr_lines = Parser::get_regular_expr_lines(rules);
+    vector<pair<string, vector<string>>> exprs = Parser::parse_expr(expr_lines, defs);
     exprs = Parser::convert_exprs_to_pos(exprs);
 
     unordered_map<string, int> priority = getPriority(exprs, keywords, punctuations);
@@ -42,24 +41,27 @@ int main() {
 
     unordered_map<int, DFA::State> dfa_states = DFA::constructDFA(exprs_nfa, priority);
     unordered_map<int, DFA::State> modified_dfa_states = DFA::processTransitions(dfa_states);
-    // Open a file for writing
 
+    // Open a file for writing
     ofstream outFile(outPath + "dfa_states_output.txt");
 
     // Check if the file is open
-    if (!outFile.is_open()) {
+    if (!outFile.is_open())
+    {
         cerr << "Error opening file for writing." << endl;
         return 1;
     }
 
     // Iterate through the DFA states map and write to the file
-    for (const auto &entry: modified_dfa_states) {
+    for (const auto &entry : modified_dfa_states)
+    {
         int dfa_state_id = entry.first;
         const DFA::State &dfa_state = entry.second;
         outFile << "DFA State ID: " << dfa_state_id << "\n";
         outFile << "NFA States: ";
 
-        for (const int &nfa_state: dfa_state.nfa_states) {
+        for (const int &nfa_state : dfa_state.nfa_states)
+        {
             outFile << nfa_state << " ";
         }
 
@@ -67,11 +69,13 @@ int main() {
         outFile << "Is Acceptance: " << (dfa_state.is_acceptance ? "true" : "false") << "\n";
         outFile << "Token: " << dfa_state.token << "\n";
         outFile << "Transitions:\n";
-        for (const auto &transition: dfa_state.transitions) {
+        for (const auto &transition : dfa_state.transitions)
+        {
             const string &input_symbol = transition.first;
             const set<int> &target_states = transition.second;
             outFile << "  " << input_symbol << " -> ";
-            for (const int &target_state: target_states) {
+            for (const int &target_state : target_states)
+            {
                 outFile << target_state << " ";
             }
             outFile << "\n";
@@ -91,19 +95,22 @@ int main() {
     ofstream outFile2(outPath + "dfa_states_output2.txt");
 
     // Check if the file is open
-    if (!outFile2.is_open()) {
+    if (!outFile2.is_open())
+    {
         cerr << "Error opening file for writing." << endl;
         return 1;
     }
 
     // Iterate through the DFA states map and write to the file
-    for (const auto &entry: minimzed_dfa_states) {
+    for (const auto &entry : minimzed_dfa_states)
+    {
         int dfa_state_id = entry.first;
         const DFA::State &dfa_state = entry.second;
 
         outFile2 << "DFA State Minimzed ID: " << dfa_state_id << "\n";
         outFile2 << "DFA States: ";
-        for (const int &nfa_state: dfa_state.nfa_states) {
+        for (const int &nfa_state : dfa_state.nfa_states)
+        {
             outFile2 << nfa_state << " ";
         }
         outFile2 << "\n";
@@ -111,12 +118,14 @@ int main() {
         outFile2 << "Token: " << dfa_state.token << "\n";
 
         outFile2 << "Transitions:\n";
-        for (const auto &transition: dfa_state.transitions) {
+        for (const auto &transition : dfa_state.transitions)
+        {
             const string &input_symbol = transition.first;
             const set<int> &target_states = transition.second;
 
             outFile2 << "  " << input_symbol << " -> ";
-            for (const int &target_state: target_states) {
+            for (const int &target_state : target_states)
+            {
                 outFile2 << target_state << " ";
             }
             outFile2 << "\n";
@@ -131,14 +140,15 @@ int main() {
     PatternMatcher pm(minimzed_dfa_states);
 
     string testProgram = "int sum , count , pass , mnt; while (pass !\\=\n"
-                        "10)\n"
-                        "{\n"
-                        "pass = pass \\+ 1 ;\n"
-                        "}";
+                         "10)\n"
+                         "{\n"
+                         "pass = pass \\+ 1 ;\n"
+                         "}";
 
     vector<pair<string, string>> symbolTable = pm.matchExpression(testProgram);
 
-    for (const auto &pair: symbolTable) {
+    for (const auto &pair : symbolTable)
+    {
         cout << "TOKEN = " << pair.first << " --- MATCHED PATTERN = " << pair.second << endl;
     }
 
@@ -151,7 +161,8 @@ int main() {
 
     symbolTable = pm.matchExpression(testString);
 
-    for (const auto &pair: symbolTable) {
+    for (const auto &pair : symbolTable)
+    {
         cout << "TOKEN = " << pair.first << " --- MATCHED PATTERN = " << pair.second << endl;
     }
 
@@ -159,7 +170,8 @@ int main() {
 
     symbolTable = pm.matchExpression(test2);
 
-    for (const auto &pair: symbolTable) {
+    for (const auto &pair : symbolTable)
+    {
         cout << "TOKEN = " << pair.first << " --- MATCHED PATTERN = " << pair.second << endl;
     }
 
@@ -167,7 +179,8 @@ int main() {
 
     symbolTable = pm.matchExpression(test3);
 
-    for (const auto &pair: symbolTable) {
+    for (const auto &pair : symbolTable)
+    {
         cout << "TOKEN = " << pair.first << " --- MATCHED PATTERN = " << pair.second << endl;
     }
 
@@ -176,7 +189,8 @@ int main() {
 
     symbolTable = pm.matchExpression(test4);
 
-    for (const auto &pair: symbolTable) {
+    for (const auto &pair : symbolTable)
+    {
         cout << "TOKEN = " << pair.first << " --- MATCHED PATTERN = " << pair.second << endl;
     }
 
@@ -185,58 +199,78 @@ int main() {
 
 unordered_map<string, int>
 getPriority(vector<pair<string, vector<string>>> exprs, const vector<string> &keywords,
-            const vector<string> &punctuations) {
+            const vector<string> &punctuations)
+{
     unordered_map<string, int> priority;
 
-    for (const string &keyword: keywords) {
+    for (const string &keyword : keywords)
+    {
         priority[keyword] = 0;
     }
-    for (const string &punctuation: punctuations) {
+    for (const string &punctuation : punctuations)
+    {
         priority[punctuation] = 0;
     }
-    for (int i = 0; i < exprs.size(); i++) {
+    for (size_t i = 0; i < exprs.size(); i++)
+    {
         priority[exprs[i].first] = i + 1;
     }
     return priority;
 }
 
-NFA convert_exprs_postfix_to_NFA(const vector<pair<string, vector<string>>> &exprs, const vector<string> &keywords,
-                                 const vector<string> &punctuations) {
+NFA convert_exprs_postfix_to_NFA(
+    const vector<pair<string, vector<string>>> &exprs,
+    const vector<string> &keywords,
+    const vector<string> &punctuations)
+{
     NFA res;
-    for (const string &keyword: keywords) {
+    for (const string &keyword : keywords)
+    {
         res.processSymbol(keyword);
         NFA expr_nfa = res.pop();
-        for (int state_i: expr_nfa.end_states) // set token of acceptance states
+        for (int state_i : expr_nfa.end_states) // set token of acceptance states
         {
             expr_nfa.states[state_i].token = keyword;
         }
         res.push(expr_nfa);
     }
-    for (const string &punctuation: punctuations) {
+    for (const string &punctuation : punctuations)
+    {
         res.processSymbol(punctuation);
         NFA expr_nfa = res.pop();
-        for (int state_i: expr_nfa.end_states) // set token of acceptance states
+        for (int state_i : expr_nfa.end_states) // set token of acceptance states
         {
             expr_nfa.states[state_i].token = punctuation;
         }
         res.push(expr_nfa);
     }
-    for (const auto &expr: exprs) {
-        for (const string &token: expr.second) {
-            if (token == "*") {
+    for (const auto &expr : exprs)
+    {
+        for (const string &token : expr.second)
+        {
+            if (token == "*")
+            {
                 res.kleeneStar();
-            } else if (token == "+") {
+            }
+            else if (token == "+")
+            {
                 res.positiveClosure();
-            } else if (token == "|") {
+            }
+            else if (token == "|")
+            {
                 res.orOp();
-            } else if (token == ".") {
+            }
+            else if (token == ".")
+            {
                 res.concatenate();
-            } else {
+            }
+            else
+            {
                 res.processSymbol(token);
             }
         }
         NFA expr_nfa = res.pop();
-        for (int state_i: expr_nfa.end_states) // set token of acceptance states
+        for (int state_i : expr_nfa.end_states) // set token of acceptance states
         {
             if (expr_nfa.states[state_i].token.empty())
                 expr_nfa.states[state_i].token = expr.first;
